requirements.txt
aiokafka
apache-airflow
numpy
minio
opencv-contrib-python
torch==1.8.0
torchvision==0.9.0
timm
mlflow
pytorch_lightning
albumentations
torchmetrics
